+++ src/Classifiers/GaussianNB.php
@@ -24,6 +24,9 @@ use Rubix\ML\Exceptions\RuntimeException;
 
 use function Rubix\ML\argmax;
 use function Rubix\ML\logsumexp;
+use function is_null;
+use function log;
+use function exp;
 
 use const Rubix\ML\TWO_PI;
 use const Rubix\ML\EPSILON;
@@ -64,6 +67,13 @@ class GaussianNB implements Estimator, Learner, Online, Probabilistic, Persistab
     protected $fitPriors;
 
     /**
+     * The amount of (epsilon) smoothing added to the variance of each feature.
+     *
+     * @var float
+     */
+    protected $smoothing;
+
+    /**
      * The weight of each class as a proportion of the entire training set.
      *
      * @var float[]
@@ -91,10 +101,18 @@ class GaussianNB implements Estimator, Learner, Online, Probabilistic, Persistab
     ];
 
     /**
-     * @param (int|float)[]|null $priors
+     * A small portion of variance to add for smoothing.
+     *
+     * @var float
+     */
+    protected $epsilon;
+
+    /**
+     * @param float[]|null $priors
+     * @param float $smoothing
      * @throws \Rubix\ML\Exceptions\InvalidArgumentException
      */
-    public function __construct(?array $priors = null)
+    public function __construct(?array $priors = null, float $smoothing = 1e-8)
     {
         $logPriors = [];
 
@@ -116,8 +134,14 @@ class GaussianNB implements Estimator, Learner, Online, Probabilistic, Persistab
             }
         }
 
+        if ($smoothing <= 0.0) {
+            throw new InvalidArgumentException('Smoothing must be'
+                . " greater than 0, $smoothing given.");
+        }
+
         $this->logPriors = $logPriors;
         $this->fitPriors = is_null($priors);
+        $this->smoothing = $smoothing;
     }
 
     /**
@@ -157,6 +181,7 @@ class GaussianNB implements Estimator, Learner, Online, Probabilistic, Persistab
     {
         return [
             'priors' => $this->fitPriors ? null : $this->priors(),
+            'smoothing' => $this->smoothing,
         ];
     }
 
@@ -226,11 +251,17 @@ class GaussianNB implements Estimator, Learner, Online, Probabilistic, Persistab
             new LabelsAreCompatibleWithLearner($dataset, $this),
         ])->check();
 
+        $maxVariance = 0.0;
+
         foreach ($dataset->stratify() as $class => $stratum) {
             if (isset($this->means[$class])) {
-                $means = $oldMeans = $this->means[$class];
-                $variances = $oldVariances = $this->variances[$class];
+                $oldMeans = $this->means[$class];
+                $oldVariances = $this->variances[$class];
                 $oldWeight = $this->weights[$class];
+
+                foreach ($oldVariances as &$oldVariance) {
+                    $oldVariance -= $this->epsilon;
+                }
 
                 $n = $stratum->numRows();
 
@@ -243,13 +274,11 @@ class GaussianNB implements Estimator, Learner, Online, Probabilistic, Persistab
                         + ($oldWeight * $oldMeans[$column]))
                         / ($oldWeight + $n);
 
-                    $vHat = ($oldWeight
+                    $variances[] = ($oldWeight
                         * $oldVariances[$column] + ($n * $variance)
                         + ($oldWeight / ($n * ($oldWeight + $n)))
                         * ($n * $oldMeans[$column] - $n * $mean) ** 2)
                         / ($oldWeight + $n);
-
-                    $variances[] = $vHat ?: EPSILON;
                 }
 
                 $weight = $oldWeight + $n;
@@ -260,11 +289,13 @@ class GaussianNB implements Estimator, Learner, Online, Probabilistic, Persistab
                     [$mean, $variance] = Stats::meanVar($values);
 
                     $means[] = $mean;
-                    $variances[] = $variance ?: EPSILON;
+                    $variances[] = $variance;
                 }
 
                 $weight = $stratum->numRows();
             }
+
+            $maxVariance = max($maxVariance, ...$variances);
 
             $this->means[$class] = $means;
             $this->variances[$class] = $variances;
@@ -278,11 +309,20 @@ class GaussianNB implements Estimator, Learner, Online, Probabilistic, Persistab
                 $this->logPriors[$class] = log($weight / $total);
             }
         }
+
+        $epsilon = $this->smoothing * ($maxVariance ?: EPSILON);
+
+        foreach ($this->variances as &$variances) {
+            foreach ($variances as &$variance) {
+                $variance += $epsilon;
+            }
+        }
+
+        $this->epsilon = $epsilon;
     }
 
     /**
-     * Calculate the likelihood of the sample being a member of a class and
-     * choose the class with the highest likelihood as the prediction.
+     * Calculate the likelihood of the sample being a member of a class and choose the class with the highest likelihood as the prediction.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      * @throws \Rubix\ML\Exceptions\RuntimeException